requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-storage
python-dotenv
openai
pydantic
httpx 